pso solver completed

+++ Solutions/base_solution.h
@@ -16,7 +16,7 @@ class BaseSolution {
 protected:
     Graph *graph;
     vector<Charger*> *charger_placement;
-    int getEvaluationSum(vector<Charger*> *charger_placement, ChanceEvaluator *chance_evaluator, EnergyEvaluator *energy_evaluator, PowerEvaluator *power_evaluator);
+    double getEvaluationSum(vector<Charger*> *charger_placement, ChanceEvaluator *chance_evaluator, EnergyEvaluator *energy_evaluator, PowerEvaluator *power_evaluator);
 public:
 
     explicit BaseSolution(Graph *graph);
